requirements.txt
redis
pydantic
python-telegram-bot